BENCH: port global optimizer benchmarks to asv
Also, remove anneal(), which is not there any more
