+++ benchmarks/benchmarks/go_benchmark_functions/__init__.py
