+++ benchmarks/benchmarks/go_benchmark_functions/go_benchmark.py
@@ -6,7 +6,10 @@ from numpy import (abs, arctan2, asarray, cos, exp, floor, log, log10,
                    arange, pi, prod, roll, seterr, sign, sin, sqrt, sum, where,
                    zeros, tan, tanh, dot)
 
-from scipy.misc import factorial
+try:
+    from scipy.misc import factorial
+except ImportError:
+    pass
 
 
 class Benchmark(object):
