+++ benchmarks/benchmarks/go_benchmark_functions/go_funcs_Y.py
