+++ benchmarks/benchmarks/go_benchmark_functions/go_funcs_univariate.py
@@ -5,7 +5,10 @@ from numpy import (abs, arctan2, asarray, cos, exp, floor, log, log10,
                    arange, pi, prod, roll, sign, sin, sqrt, sum, where,
                    zeros, tan, tanh, dot)
 
-from scipy.misc import factorial
+try:
+    from scipy.misc import factorial
+except ImportError:
+    pass
 from .go_benchmark import Benchmark
 
 #-----------------------------------------------------------------------
