+++ benchmarks/benchmarks/optimize.py
@@ -1,7 +1,8 @@
 from __future__ import division, print_function, absolute_import
 
 import time
-from collections import defaultdict
+import inspect
+from collections import defaultdict, OrderedDict
 
 import numpy as np
 
@@ -9,10 +10,13 @@ try:
     import scipy.optimize
     from scipy.optimize.optimize import rosen, rosen_der, rosen_hess
     from scipy.optimize import leastsq
+    from scipy.optimize import (leastsq, basinhopping, differential_evolution,
+                                OptimizeResult)
 except ImportError:
     pass
 
 from . import test_functions as funcs
+from . import go_benchmark_functions as gbf
 from .common import Benchmark
 from .lsq_problems import extract_lsq_problems
 
@@ -43,8 +47,27 @@ class _BenchOptimizers(Benchmark):
 
         self.results = []
 
+    @classmethod
+    def from_funcobj(cls, function_name, function, **minimizer_kwargs):
+        self = cls.__new__(cls)
+        self.function_name = function_name
+
+        self.function = function
+        self.fun = function.fun
+        if hasattr(function, 'der'):
+            self.der = function.der
+
+        self.bounds = function.bounds
+
+        self.minimizer_kwargs = minimizer_kwargs
+        self.results = []
+        return self
+
     def reset(self):
         self.results = []
+
+    def energy_gradient(self, x):
+        return self.fun(x), self.function.der(x)
 
     def add_result(self, result, t, name):
         """add a result to the list"""
@@ -89,12 +112,93 @@ class _BenchOptimizers(Benchmark):
             newres.mean_time = np.mean([r.time for r in result_list])
             newres.ntrials = len(result_list)
             newres.nfail = len([r for r in result_list if not r.success])
+            newres.nsuccess = len([r for r in result_list if r.success])
             try:
                 newres.ndim = len(result_list[0].x)
             except TypeError:
                 newres.ndim = 1
             averaged_results[name] = newres
-        return averaged_results.values()
+        return averaged_results
+
+    # for basinhopping
+    def accept_test(self, x_new=None, *args, **kwargs):
+        """
+        Does the new candidate vector lie inbetween the bounds?
+
+        Returns
+        -------
+        accept_test : bool
+            The candidate vector lies inbetween the bounds
+        """
+        if not hasattr(self.function, "xmin"):
+            return True
+        if np.any(x_new < self.function.xmin):
+            return False
+        if np.any(x_new > self.function.xmax):
+            return False
+        return True
+
+    def run_basinhopping(self):
+        """
+        Do an optimization run for basinhopping
+        """
+        kwargs = self.minimizer_kwargs
+        if hasattr(self.fun, "temperature"):
+            kwargs["T"] = self.function.temperature
+        if hasattr(self.fun, "stepsize"):
+            kwargs["stepsize"] = self.function.stepsize
+
+        minimizer_kwargs = {"method": "L-BFGS-B"}
+
+        x0 = self.function.initial_vector()
+
+        # basinhopping - no gradient
+        minimizer_kwargs['jac'] = False
+        self.function.nfev = 0
+
+        t0 = time.time()
+
+        res = basinhopping(
+            self.fun, x0, accept_test=self.accept_test,
+            minimizer_kwargs=minimizer_kwargs,
+            **kwargs)
+
+        t1 = time.time()
+        res.success = self.function.success(res.x)
+        res.nfev = self.function.nfev
+        self.add_result(res, t1 - t0, 'basinh.')
+
+    def run_differentialevolution(self):
+        """
+        Do an optimization run for differential_evolution
+        """
+        self.function.nfev = 0
+
+        t0 = time.time()
+
+        res = differential_evolution(self.fun,
+                                     self.bounds,
+                                     popsize=20)
+
+        t1 = time.time()
+        res.success = self.function.success(res.x)
+        res.nfev = self.function.nfev
+        self.add_result(res, t1 - t0, 'DE')
+
+    def bench_run_global(self, numtrials=50, methods=None):
+        """
+        Run the optimization tests for the required minimizers.
+        """
+
+        if methods is None:
+            methods = ['DE', 'basinh.']
+
+        method_fun = {'DE': self.run_differentialevolution,
+                      'basinh.': self.run_basinhopping}
+
+        for i in range(numtrials):
+            for m in methods:
+                method_fun[m]()
 
     def bench_run(self, x0, methods=None, **minimizer_kwargs):
         """do an optimization test starting at x0 for all the optimizers"""
@@ -154,15 +258,10 @@ class BenchSmoothUnbounded(Benchmark):
 
     def setup(self, func_name, method_name, ret_val):
         b = getattr(self, 'run_' + func_name)(methods=[method_name])
-        results = b.average_results()
-        result = None
-        for r in results:
-            if r.name == method_name:
-                result = getattr(r, ret_val)
-                break
-        if result is None:
+        r = b.average_results().get(method_name)
+        if r is None:
             raise NotImplementedError()
-        self.result = result
+        self.result = getattr(r, ret_val)
 
     def track_all(self, func_name, method_name, ret_val):
         return self.result
@@ -276,3 +375,39 @@ class BenchLeastSquares(Benchmark):
             return int(problem.check_answer(x, ftol))
         else:
             raise NotImplementedError
+
+
+class BenchGlobal(Benchmark):
+    """
+    Benchmark the global optimizers using the go_benchmark_functions
+    suite
+    """
+    _functions = OrderedDict([
+        item for item in inspect.getmembers(gbf, inspect.isclass)
+        if (issubclass(item[1], gbf.Benchmark) and
+            item[0] not in ('Benchmark', 'LennardJones') and
+            not item[0].startswith('Problem'))
+    ])
+
+    params = [
+        _functions.keys(),
+        ["success%", "<nfev>"],
+        ['DE', 'basinh.'],
+    ]
+    param_names = ["test function", "result type", "solver"]
+
+    def track_all(self, name, ret_value, solver):
+        klass = self._functions[name]
+        numtrials = 1
+
+        f = klass()
+        b = _BenchOptimizers.from_funcobj(name, f)
+        b.bench_run_global(methods=[solver], numtrials=numtrials)
+        av_results = b.average_results()
+
+        if ret_value == 'success%':
+            return 100 * av_results[solver].nsuccess / numtrials
+        elif ret_value == '<nfev>':
+            return av_results[solver].mean_nfev
+        else:
+            raise ValueError()
